+++ editor/src/components/canvas/canvas-strategies/strategies/flex-resize-strategy.tsx
@@ -64,7 +64,7 @@ import { strictEvery, mapDropNulls } from '../../../../core/shared/array-utils'
 import type { ElementPath } from '../../../../core/shared/project-file-types'
 import type { ElementPathTrees } from '../../../../core/shared/element-path-tree'
 import { treatElementAsGroupLike } from './group-helpers'
-import { queueTrueUpElement } from '../../commands/queue-group-true-up-command'
+import { queueTrueUpElement } from '../../commands/queue-true-up-command'
 import {
   pushIntendedBoundsGroup,
   pushIntendedBoundsAndUpdateTargets,
