+++ editor/src/components/editor/actions/actions.tsx
@@ -526,7 +526,7 @@ import {
 } from '../../canvas/canvas-strategies/strategies/group-conversion-helpers'
 import { addElements } from '../../canvas/commands/add-elements-command'
 import { deleteElement } from '../../canvas/commands/delete-element-command'
-import { queueTrueUpElement } from '../../canvas/commands/queue-group-true-up-command'
+import { queueTrueUpElement } from '../../canvas/commands/queue-true-up-command'
 import { processWorkerUpdates } from '../../../core/shared/parser-projectcontents-utils'
 import { getAllUniqueUids } from '../../../core/model/get-unique-ids'
 import { getLayoutProperty } from '../../../core/layout/getLayoutProperty'
