+++ editor/src/components/inspector/inspector-strategies/fixed-edge-basic-strategy.ts
@@ -7,7 +7,7 @@ import {
 } from '../../canvas/commands/set-css-length-command'
 import type { CSSNumber } from '../common/css-utils'
 import type { InspectorStrategy } from './inspector-strategy'
-import { queueTrueUpElement } from '../../canvas/commands/queue-group-true-up-command'
+import { queueTrueUpElement } from '../../canvas/commands/queue-true-up-command'
 import {
   groupErrorToastCommand,
   maybeInvalidGroupState,
