+++ editor/src/components/inspector/inspector-strategies/fixed-size-basic-strategy.ts
@@ -10,7 +10,7 @@ import type { CSSNumber } from '../common/css-utils'
 import type { Axis } from '../inspector-common'
 import { removeExtraPinsWhenSettingSize, widthHeightFromAxis } from '../inspector-common'
 import type { InspectorStrategy } from './inspector-strategy'
-import { queueTrueUpElement } from '../../canvas/commands/queue-group-true-up-command'
+import { queueTrueUpElement } from '../../canvas/commands/queue-true-up-command'
 import {
   groupErrorToastCommand,
   maybeInvalidGroupState,
