+++ editor/src/components/inspector/inspector-strategies/hug-contents-basic-strategy.ts
@@ -25,7 +25,7 @@ import {
   widthHeightFromAxis,
 } from '../inspector-common'
 import type { InspectorStrategy } from './inspector-strategy'
-import { queueTrueUpElement } from '../../canvas/commands/queue-group-true-up-command'
+import { queueTrueUpElement } from '../../canvas/commands/queue-true-up-command'
 import { trueUpGroupElementChanged } from '../../../components/editor/store/editor-state'
 
 const CHILDREN_CONVERTED_TOAST_ID = 'CHILDREN_CONVERTED_TOAST_ID'
